backend/requirements.txt
fastapi
uvicorn
chromadb
openai
PyMuPDF
python-multipart
python-dotenv
